Refactor Asynctask and "ping" webserver
To detect if the direction exists and is running Yoke, the app downloads a
lightweight file. In this way, it can display an error message if it fails.

This reduces frustration due to user error when Yoke is not running or the
wi-fi is disabled. No, I wouldn't know by experience, why do you ask.