Yoke now warns if the user deliberately closes the connection
This simplifies some checks in service.py. It should also allow for
users to come back to the app to recover the connection if they
accidentally exited the app, but this requires more tests and raising
the timeout limit, so nothing has been done in this direction yet.

An important change was made to accomodate this: THE YOKE APP NOW
ACTUALLY ENFORCES A CONSISTENT POLL RATE. (Previously it enforced only
a minimum poll rate.) This can be lowered or made customizable in the
future.